# Conflicts:
#	Trust.xcodeproj/project.pbxproj
#	Trust/Core/Initializers/MirgrationInitializer.swift

+++ Trust/Models/SignTransaction.swift
@@ -2,13 +2,14 @@
 
 import Foundation
 import BigInt
+import TrustCore
 import TrustKeystore
 
 public struct SignTransaction {
     let value: BigInt
     let account: Account
     let to: Address?
-    let nonce: Int
+    let nonce: BigInt
     let data: Data
     let gasPrice: BigInt
     let gasLimit: BigInt
